+++ Frontend/package.json
@@ -41,6 +41,7 @@
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^15.15.0",
+    "path": "^0.12.7",
     "vite": "^6.3.3"
   }
 }
