+++ Frontend/vite.config.js
@@ -1,7 +1,13 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import tailwindcss from '@tailwindcss/vite'
-// https://vite.dev/config/
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import tailwindcss from '@tailwindcss/vite';
+import path from 'path'; // 👉 เพิ่มบรรทัดนี้
+
 export default defineConfig({
-  plugins: [react(),tailwindcss()],
-})
+  plugins: [react(), tailwindcss()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, 'src'), // 👉 เพิ่ม alias
+    },
+  },
+});
